backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.2.0
dj-database-url==2.0.0
Wand==0.6.13
python-decouple==3.6
psycopg2-binary==2.9.11
setuptools==68.0.0
gunicorn==20.1.0
